requirements.txt
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
Beaker==1.12.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
executing==1.2.0
Flask==2.2.2
Flask-Login==0.6.2
flask-redis==0.4.0
Flask-SQLAlchemy==3.0.2
future==0.18.2
greenlet==2.0.1
idna==3.4
importlib-metadata==6.0.0
ipython==8.8.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jwcrypto==1.4.2
Mako==1.2.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
oic==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycryptodomex==3.16.0
Pygments==2.14.0
pyjwkest==1.4.2
PyJWT==2.6.0
python-dotenv==0.21.0
redis==4.4.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.46
stack-data==0.6.2
traitlets==5.8.1
typing_extensions==4.4.0
urllib3==1.26.13
uWSGI==2.0.21
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
